Add crop_canvas, substitute canvas data type
Add crop_canvas method to chop off the rows and columns that are below the visual borders of the painted canvas. Substitute canvas to be a numpy array instead of a list of lists; the reason being, faster operations on a multidimensional array, which is helpful at the very least on a crop. Clean up the paint_canvas method and the corresponding tests with cropping-related crutches.