Add error message propagation, fix csv-related issues
Also added general error handling and did some refactoring. The non-ascii symbols for comments and delimiters in csv's are exactly what you think they are for -- so that the loadtxt doesn't misinterpret the actual ascii characters as them.
Previously, a # was used for comments and that broke loading any csv with rectangles with that symbol.